+++ R/rmw_do_all.R
@@ -54,11 +54,16 @@
 #' 
 #' \donttest{
 #' 
+#' # Load package
+#' library(dplyr)
+#' 
 #' # Keep things reproducible
 #' set.seed(123)
 #' 
 #' # Prepare example data
-#' data_london_prepared <- rmw_prepare_data(data_london, value = "no2")
+#' data_london_prepared <- data_london %>% 
+#'   filter(variable == "no2") %>% 
+#'   rmw_prepare_data()
 #' 
 #' # Use the example data to conduct the steps needed for meteorological
 #' # normalisation