+++ R/rmw_predict.R
@@ -18,17 +18,25 @@
 #' 
 #' @examples 
 #' 
+#' # Load package
+#' library(dplyr)
+#' 
+#' # Prepare example data
+#' data_london_prepared <- data_london %>% 
+#'   filter(variable == "no2") %>% 
+#'   rmw_prepare_data()
+#' 
 #' # Make a prediction with the examples
 #' vector_prediction <- rmw_predict(
 #'   model_london, 
-#'   df = rmw_prepare_data(data_london, value = "no2")
+#'   df = data_london_prepared
 #' )
 #' 
 #' 
 #' # Make a prediction with standard errors too
 #' list_prediction <- rmw_predict(
 #'   model_london, 
-#'   df = rmw_prepare_data(data_london, value = "no2"),
+#'   df = data_london_prepared,
 #'   se = TRUE
 #' )
 #'