+++ tests/testthat.R
@@ -1,4 +1,5 @@
 library(testthat)
+library(dplyr)
 library(rmweather)
 
 test_check("rmweather")