+++ NEWS.md
@@ -1,3 +1,9 @@
+# rmweather 0.1.5
+
+  - Minor changes to improve compatibility with dplyr 1.0.0
+  
+  - Update package's data and alter format where pollutants are in a "long" format
+
 # rmweather 0.1.4
 
   - Replace deprecated `data_frame` with `tibble`